address PR review + /simplify: require KOLU_GH_BIN; consolidate button
PR review (srid @ nix/env.nix:13): removed the `?? "gh"` fallback in
`github.ts` and made KOLU_GH_BIN required — throws at module load if
unset. The dev shell (via `shell.nix` importing `koluEnv`) and the
packaged wrapper (now via `--set KOLU_GH_BIN` in `default.nix`) both
set it, so Nix-run kolu always has a pinned gh path. Non-Nix runs are
unsupported; failing fast is clearer than silently resolving a
different `gh` from PATH.

/simplify findings:
- Swap inline navigator.clipboard.writeText for writeTextToClipboard
  (packages/client/src/terminal/clipboard.ts) so the copy button keeps
  working in non-secure contexts via the execCommand fallback.
- Move PrUnavailableButton out of TerminalMeta.tsx and into
  PrUnavailablePopover.tsx (exported). The button + popover + per-site
  open-state signal are one unit; hosting them in the same module
  lets TerminalMeta.tsx stay a thin render site.
- Trim redundant JSDoc in PrUnavailablePopover.tsx (the file header
  now carries the "why two-layer dispatch + why Portal not Corvu"
  note; per-component doc comments that restated it are gone).
