+++ nix/env.nix
@@ -9,8 +9,9 @@
 {
   KOLU_FONTS_DIR          = pkgs.kolu-fonts;
   KOLU_CLIPBOARD_SHIM_DIR = "${pkgs.kolu-clipboard-shims}/bin";
-  # Pinned gh binary — the server's GitHub provider reads this and falls
-  # back to PATH lookup when unset (dev shells, non-Nix installs). See
-  # packages/server/src/meta/github.ts for the read site.
+  # Pinned gh binary — the server's GitHub provider consumes this directly.
+  # Required, not optional: github.ts throws at startup if unset. Set here so
+  # both the packaged wrapper (default.nix) and the dev shell (shell.nix)
+  # pick it up via `koluEnv`.
   KOLU_GH_BIN             = "${pkgs.gh}/bin/gh";
 }
